Refactor route changes
