--- app/client/src/ce/pages/Editor/AdditionalEditorRoutes.tsx
@@ -1,9 +1,0 @@
-interface Props {
-  path: string;
-}
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function AdditionalEditorRoutes(props: Props) {
-  return null;
-}
-
-export default AdditionalEditorRoutes;
