--- app/client/src/ee/pages/Editor/AdditionalEditorRoutes.tsx
@@ -1,11 +1,0 @@
-export * from "ce/pages/Editor/AdditionalEditorRoutes";
-
-interface Props {
-  path: string;
-}
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function AdditionalEditorRoutes(props: Props) {
-  return null;
-}
-
-export default AdditionalEditorRoutes;
